Fix RecordIn and RecordDrop stats
Summary:
1. fix possible overflow of the two stats by using uint64_t
2. use a similar source of data to calculate RecordDrop. Previous one is not correct.

Test Plan: See outputs of db_bench settings, and the results look reasonable

Reviewers: MarkCallaghan, ljin, igor

Reviewed By: igor

Subscribers: rven, leveldb, yhchiang, dhruba

Differential Revision: https://reviews.facebook.net/D28155

+++ db/internal_stats.h
@@ -127,11 +127,11 @@ class InternalStats {
     int files_out_levelnp1;
 
     // Total incoming entries during compaction between levels N and N+1
-    int num_input_records;
+    uint64_t num_input_records;
 
     // Accumulated diff number of entries
     // (num input entries - num output entires) for compaction  levels N and N+1
-    int num_dropped_records;
+    uint64_t num_dropped_records;
 
     // Number of compactions done
     int count;
